% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixNB.R
\name{DeMixNB}
\alias{DeMixNB}
\title{Deconvolution of heterogeneous tumor samples with two components using
expression data from RNAseq or microarray platforms}
\usage{
DeMixNB(
  data.Y,
  data.N1,
  niter = 30,
  nspikein = 200,
  tol = 1e-08,
  nthread = 1,
  seeds = c(123, 456, 789),
  output.more.info = FALSE,
  consistency_threshold = 0.1
)
}
\arguments{
\item{data.Y}{A SummarizedExperiment object of expression data from mixed 
tumor samples. It is a \eqn{G} by \eqn{My} matrix where \eqn{G} is the number
of genes and \eqn{My} is the number of mixed samples. Samples with the same
tissue type should be placed together in columns.}

\item{data.N1}{A SummarizedExperiment object of expression data 
from reference component 1 (e.g., normal). It is a \eqn{G} by \eqn{M1} matrix 
where \eqn{G} is the number of genes and \eqn{M1} is the number of samples 
for component 1.}

\item{niter}{The maximum number of iterations used in the algorithm of 
iterated conditional modes. A larger value better guarantees 
the convergence in estimation but increases the running time. The default is 
30.}

\item{nspikein}{The number of spikes in normal reference used for proportion
estimation. The default value is \eqn{ min(200, My)}, where 
\eqn{My} the number of mixed samples. If it is set to 0, proportion 
estimation is performed without any spike in normal reference.}

\item{tol}{The convergence criterion. The default is 10^(-8).}

\item{nthread}{The number of cores to use for parallelization. The default
is 1, i.e. no parallelization at the top level. Note that OpenMP is used 
under the hood when calculating the complete likelihood.}

\item{seeds}{A numeric vector of random seeds to use for each run. 
The DeMixNB algorithm will be run once for each seed. Default is c(123, 456, 789).}

\item{output.more.info}{A boolean flag controlling whether to return the 
estimations in each iteration. Default is False.}

\item{consistency_threshold}{The maximum allowed difference (range) 
between pi estimates across runs for a spot to be considered "consistent". 
Default is 0.1.}
}
\value{
\item{pi_t_summary}{A data frame containing the estimated tumor proportions for
each run and a consistency flag at each sample indicating whether the runs agree 
with each other based on the maximum difference.}

\item{pi_n_summary}{A data frame containing the estimated non-tumor proportions 
for each run.}

\item{all_run_results}{A list containing all the detailed outputs of each run,
including the \eqn{mu} and \eqn{phi} estimates.}
}
\description{
DeMixNB is a variant of DeMixT such that the expressions are 
assumed to follow the negative binomial function
}
\examples{

# Example: estimate proportions for simulated two-component data 
  data(test.data.2comp)
# res.NB = DeMixNB(data.Y = test.data.2comp$data.Y, 
#                    data.N1 = test.data.2comp$data.N1,
#                    niter = 10, tol = 10^(-5))

}
\author{
Liyang Xie
}
