% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixNB.R
\name{DeMixNB_single_seed}
\alias{DeMixNB_single_seed}
\title{Deconvolution of heterogeneous tumor samples with two components using
expression data from RNAseq or microarray platforms by single seed}
\usage{
DeMixNB_single_seed(
  data.Y,
  data.N1,
  niter = 30,
  nspikein = 200,
  tol = 1e-07,
  nthread = 1,
  output.more.info = FALSE,
  seed = 100
)
}
\arguments{
\item{data.Y}{A SummarizedExperiment object of expression data from mixed 
tumor samples. It is a \eqn{G} by \eqn{My} matrix where \eqn{G} is the number
of genes and \eqn{My} is the number of mixed samples. Samples with the same
tissue type should be placed together in columns.}

\item{data.N1}{A SummarizedExperiment object of expression data 
from reference component 1 (e.g., normal). It is a \eqn{G} by \eqn{M1} matrix 
where \eqn{G} is the number of genes and \eqn{M1} is the number of samples 
for component 1.}

\item{niter}{The maximum number of iterations used in the algorithm of 
iterated conditional modes. A larger value better guarantees 
the convergence in estimation but increases the running time. The default is 
30.}

\item{nspikein}{The number of spikes in normal reference used for proportion
estimation. The default value is \eqn{ min(200, My)}, where 
\eqn{My} the number of mixed samples. If it is set to 0, proportion 
estimation is performed without any spike in normal reference.}

\item{tol}{The convergence criterion. The default is 10^(-8).}

\item{nthread}{The number of cores to use for parallelization. The default
is 1, i.e. no parallelization at the top level. Note that OpenMP is used 
under the hood when calculating the complete likelihood.}

\item{output.more.info}{A boolean flag controlling whether to return the 
estimations in each iteration. Default is False.}

\item{seed}{The random seed to use for each run.}
}
\value{
\item{pi}{A matrix of estimated proportion. First row corresponds to the 
proportion estimate for the known components and unknown component 
respectively, and each column 
corresponds to one sample.}

\item{pi.iter}{Estimated proportions in each iteration. It is a \eqn{niter*
My*2} array. This is enabled only when output.more.info = TRUE.}

\item{Mu}{A matrix of estimated \eqn{mu} of the negative binomial 
distribution for both components. Each row corresponds to one gene.} 
\item{Phi}{A matrix of estimated \eqn{phi} of the negative binomial 
distribution for both components. Each row corresponds to one gene.}
\item{gene.name}{The names of genes used in estimating the proportions. 
If no gene names are provided in the original data set, the genes will be
automatically indexed.}
}
\description{
The DeMixNB algorithm with a single seed.
}
