\name{test.data1.truth}
\alias{test.data1.truth}
\docType{data}
\title{simulated two-component test data}

\description{simulated two-component test data used in function DeMixT}

\usage{test.data1.truth}

\format{
A SummarizedExperiment object of true proportions, i.e., \eqn{\pi_1} and
\eqn{1 - \pi_1}
}

\examples{
data(test.data1.truth)
test.data1.truth
}
\keyword{datasets}