\name{test.data1.y}
\alias{test.data1.y}
\docType{data}
\title{simulated two-component test data}

\description{simulated two-component test data used in function DeMixT}

\usage{test.data1.y}

\format{
A SummarizedExperiment object of expression data from mixed tumor samples
}

\examples{
data(test.data1.y)
test.data1.y
}
\keyword{datasets}