\name{test.data2.comp1}
\alias{test.data2.comp1}
\docType{data}
\title{simulated three-component test data}

\description{simulated three-component test data used in function DeMixT}

\usage{test.data2.comp1}

\format{
A SummarizedExperiment object of expression data from reference component 1
}

\examples{
data(test.data2.comp1)
test.data2.comp1
}
\keyword{datasets}