\name{test.data2.comp2}
\alias{test.data2.comp2}
\docType{data}
\title{simulated three-component test data}

\description{simulated three-component test data used in function DeMixT}

\usage{test.data2.comp2}

\format{
A SummarizedExperiment object of expression data from reference component 2
}

\examples{
data(test.data2.comp2)
test.data2.comp2
}

\keyword{datasets}