\name{test.data2.truth}
\alias{test.data2.truth}
\docType{data}
\title{simulated three-component test data}

\description{simulated three-component test data used in function DeMixT}

\usage{test.data2.truth}

\format{
A SummarizedExperiment object of true proportions, i.e., \eqn{\pi_1} 
and  \eqn{1 - \pi_1}
}

\examples{
data(test.data2.truth)
test.data2.truth
}
\keyword{datasets}