\name{test.data2.y}
\alias{test.data2.y}
\docType{data}
\title{simulated three-component test data}

\description{simulated three-component test data used in function DeMixT}

\usage{test.data2.y}

\format{
A SummarizedExperiment object of expression data from mixed tumor samples
}

\examples{
data(test.data2.y)
test.data2.y
}
\keyword{datasets}