\name{test.data3.comp1}
\alias{test.data3.comp1}
\docType{data}
\title{three-component mixed cell line test data}

\description{three-component mixed cell line test data used in function DeMixT}

\usage{test.data3.comp1}

\format{
A SummarizedExperiment object of expression data from reference component 1
}

\examples{
data(test.data3.comp1)
test.data3.comp1
}
\keyword{datasets}