\name{test.data3.comp2}
\alias{test.data3.comp2}
\docType{data}
\title{three-component mixed cell line test data}

\description{three-component mixed cell line test data used in function DeMixT}

\usage{test.data3.comp2}

\format{
A SummarizedExperiment object of expression data from reference component 2
}

\examples{
data(test.data3.comp2)
test.data3.comp2
}
\keyword{datasets}