\name{test.data3.truth}
\alias{test.data3.truth}
\docType{data}
\title{three-component mixed cell line test data}

\description{three-component mixed cell line test data used in function DeMixT}

\usage{test.data3.truth}

\format{
A SummarizedExperiment object of true proportions, i.e., \eqn{\pi_1} and
\eqn{1 - \pi_1}
}

\examples{
data(test.data3.truth)
test.data3.truth
}
\keyword{datasets}