\name{test.data3.y}
\alias{test.data3.y}
\docType{data}
\title{three-component mixed cell line test data}

\description{three-component mixed cell line test data used in function DeMixT}

\usage{test.data3.y}

\format{
three-component mixed cell line test data used in function DeMixT
}

\examples{
data(test.data3.y)
test.data3.y
}
\keyword{datasets}