% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixT_preprocessing.R
\name{batch_correction}
\alias{batch_correction}
\title{batch_correction}
\usage{
batch_correction(count.matrix, batch_labels)
}
\arguments{
\item{count.matrix}{A matrix of raw expression count with \eqn{G} by \eqn{(My)}, where \eqn{G} is the number
of genes, \eqn{My} is the number of mixed tumor samples. Row names are genes
 column names are tumor sample ids.}

\item{batch_labels}{Factor of tumor samples from different batches}
}
\value{
Batch effect corrected count matrix for tumor samples
}
\description{
Batch effect correction for multiple batches of tumor samples using ComBat
}
