% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixT_preprocessing.R
\name{plot_dim}
\alias{plot_dim}
\title{plot_dim}
\usage{
plot_dim(
  count.matrix,
  labels,
  legend.position = "bottomleft",
  legend.cex = 1.2
)
}
\arguments{
\item{count.matrix}{A matrix of raw expression count with \eqn{G} by \eqn{(My + M1)}, where \eqn{G} is the number
of genes, \eqn{My} is the number of mixed samples and \eqn{M1} is the number of normal samples. Row names are genes
 column names are sample ids.}

\item{labels}{A vector of sample labels for plotting}

\item{legend.position}{Position of legend in the plot. Default is bottomleft.}

\item{legend.cex}{Character expansion factor relative to current par("cex"). Default = 1.2}
}
\value{
None (create plots)
}
\description{
Plot the distribution of tumor and normal samples in a 2D PCA space based on their expressions
}
