% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeMixNB_preprocessing.R,
%   R/DeMixT_preprocessing.R
\name{scale_normalization_75th_percentile}
\alias{scale_normalization_75th_percentile}
\title{scale_normalization_75th_percentile}
\usage{
scale_normalization_75th_percentile(count.matrix)

scale_normalization_75th_percentile(count.matrix)
}
\arguments{
\item{count.matrix}{A matrix of raw expression count with \eqn{G} by \eqn{(My + M1)}, where \eqn{G} is the number
of genes, \eqn{My} is the number of mixed samples and \eqn{M1} is the number of normal samples. Row names are genes
 column names are sample ids.}
}
\value{
the scale normalized count matrix

the scale normalized count matrix
}
\description{
Quantile normalization for the raw count matrix of tumor and normal reference using the 0.75 quantile scale normalization

Quantile normalization for the raw count matrix of tumor and normal reference using the 0.75 quantile scale normalization
}
